+++ src/MicroElements.Functional/StringFormatExtensions.cs
@@ -1,7 +1,9 @@
 ﻿// Copyright (c) MicroElements. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System;
 using System.Collections.Generic;
+using System.Globalization;
 using System.Text;
 
 namespace MicroElements.Functional
@@ -11,14 +13,42 @@ namespace MicroElements.Functional
     /// </summary>
     public static class StringFormatExtensions
     {
+        public static readonly NumberFormatInfo DefaultNumberFormatInfo = NumberFormatInfo.ReadOnly(
+            new NumberFormatInfo
+            {
+                NumberDecimalSeparator = ".",
+            });
+
+        public static Func<object, string> DefaultFormatValue = value =>
+        {
+            switch (value)
+            {
+                case double num:
+                    return num.ToString(DefaultNumberFormatInfo);
+                case float num:
+                    return num.ToString(DefaultNumberFormatInfo);
+                case decimal num:
+                    return num.ToString(DefaultNumberFormatInfo);
+                case DateTime dateTime:
+                    return dateTime == dateTime.Date ? $"{dateTime:yyyy-MM-dd}" : $"{dateTime:yyyy-MM-ddTHH:mm:ss}";
+                default:
+                    return $"{value}";
+            }
+        };
+
         /// <summary>
         /// Formats enumeration of value as tuple: (value1, value2, ...).
         /// </summary>
         /// <param name="values">Values enumeration.</param>
         /// <param name="fieldSeparator">Optional field separator.</param>
         /// <param name="nullPlaceholder">Optional null placeholder.</param>
+        /// <param name="formatValue">Func to format object value to string representation.</param>
         /// <returns>Formatted string.</returns>
-        public static string FormatAsTuple(this IEnumerable<object> values, string fieldSeparator = ", ", string nullPlaceholder = "null")
+        public static string FormatAsTuple(
+            this IEnumerable<object> values,
+            string fieldSeparator = ", ",
+            string nullPlaceholder = "null",
+            Func<object, string> formatValue = null)
         {
             values.AssertArgumentNotNull(nameof(values));
             fieldSeparator.AssertArgumentNotNull(nameof(fieldSeparator));
@@ -40,22 +70,28 @@ namespace MicroElements.Functional
         /// <param name="values">Values enumeration.</param>
         /// <param name="fieldSeparator">Optional field separator.</param>
         /// <param name="nullPlaceholder">Optional null placeholder.</param>
+        /// <param name="formatValue">Func to format object value to string representation.</param>
         /// <returns>Formatted string.</returns>
-        public static string FormatAsJson(this IEnumerable<(string Name, string Value)> values, string fieldSeparator = ", ", string nullPlaceholder = "null")
+        public static string FormatAsJson(
+            this IEnumerable<(string Name, object Value)> values,
+            string fieldSeparator = ", ",
+            string nullPlaceholder = "null",
+            Func<object, string> formatValue = null)
         {
             values.AssertArgumentNotNull(nameof(values));
             fieldSeparator.AssertArgumentNotNull(nameof(fieldSeparator));
             nullPlaceholder.AssertArgumentNotNull(nameof(nullPlaceholder));
 
+            formatValue = formatValue ?? DefaultFormatValue;
+
             var stringBuilder = new StringBuilder();
             stringBuilder.Append("{");
-            foreach ((string Name, string Value) formatComponent in values)
+            foreach ((string Name, object Value) formatComponent in values)
             {
-                if (formatComponent.Value != null)
-                    stringBuilder.Append($"{formatComponent.Name}: \"{formatComponent.Value}\"{fieldSeparator}");
-                else
-                    stringBuilder.Append($"{formatComponent.Name}: {nullPlaceholder}{fieldSeparator}");
+                var formatted = formatValue(formatComponent.Value) ?? nullPlaceholder;
+                stringBuilder.Append($"{formatComponent.Name}: \"{formatted}\"{fieldSeparator}");
             }
+
             if (stringBuilder.Length > fieldSeparator.Length)
                 stringBuilder.Length -= fieldSeparator.Length;
             stringBuilder.Append("}");